+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 * Add `header` and `footer` directives ([#22](https://github.com/marp-team/marpit/pull/22))
+* Support importing other theme CSS with `@import` (or `@import-theme`) ([#24](https://github.com/marp-team/marpit/pull/24))
 
 ## v0.0.5 - 2018-05-12